+++ docs/changelog.rst
@@ -2,8 +2,14 @@
 CHANGELOG
 =========
 
-2.92.2+dev (XXXX-XX-XX)
+2.92.3     (2022-12-02)
 -----------------------
+
+**Improvements**
+
+- API v2:
+  - revert ``?trek filter`` by direct intersecting geometry on sensitive area endpoint.
+  - improve ``?near_xxx`` filters by direct intersecting buffered geometry on sensitive area endpoint.
 
 
 2.92.2     (2022-12-01)